# Move filtering logic from RawScorer to FilteredScorer