--- a/lib/segment/benches/hnsw_build_asymptotic.rs
+++ b/lib/segment/benches/hnsw_build_asymptotic.rs
@@ -3,11 +3,11 @@ mod prof;
 
 use std::cell::LazyCell;
 
+use common::iterator_ext::IteratorExt as _;
 use criterion::{Criterion, criterion_group, criterion_main};
 use itertools::Itertools;
 use rand::{Rng, rng};
-use segment::fixtures::index_fixtures::{FakeFilterContext, TestRawScorerProducer, random_vector};
-use segment::index::hnsw_index::point_scorer::FilteredScorer;
+use segment::fixtures::index_fixtures::{TestRawScorerProducer, random_vector};
 use segment::spaces::metric::Metric;
 use segment::spaces::simple::{CosineMetric, DotProductMetric};
 use segment::vector_storage::DEFAULT_STOPPED;
@@ -34,10 +34,8 @@ fn hnsw_build_asymptotic(c: &mut Criterion) {
     group.bench_function("build-n-search-hnsw-5k", |b| {
         let (vector_holder, graph_layers) = &*setup_5k;
         b.iter(|| {
-            let fake_filter_context = FakeFilterContext {};
             let query = random_vector(&mut rng, DIM);
-            let raw_scorer = vector_holder.get_raw_scorer(query).unwrap();
-            let scorer = FilteredScorer::new(raw_scorer.as_ref(), Some(&fake_filter_context));
+            let scorer = vector_holder.get_scorer(query);
             graph_layers
                 .search(TOP, EF, scorer, None, &DEFAULT_STOPPED)
                 .unwrap();
@@ -55,10 +53,8 @@ fn hnsw_build_asymptotic(c: &mut Criterion) {
     group.bench_function("build-n-search-hnsw-1M", |b| {
         let (vector_holder, graph_layers) = &*setup_1m;
         b.iter(|| {
-            let fake_filter_context = FakeFilterContext {};
             let query = random_vector(&mut rng, DIM);
-            let raw_scorer = vector_holder.get_raw_scorer(query).unwrap();
-            let scorer = FilteredScorer::new(raw_scorer.as_ref(), Some(&fake_filter_context));
+            let scorer = vector_holder.get_scorer(query);
             graph_layers
                 .search(TOP, EF, scorer, None, &DEFAULT_STOPPED)
                 .unwrap();
@@ -68,15 +64,13 @@ fn hnsw_build_asymptotic(c: &mut Criterion) {
     group.bench_function("build-n-search-hnsw-1M-score-point", |b| {
         let (vector_holder, _graph_layers) = &*setup_1m;
         b.iter(|| {
-            let fake_filter_context = FakeFilterContext {};
             let query = random_vector(&mut rng, DIM);
-            let raw_scorer = vector_holder.get_raw_scorer(query).unwrap();
-            let mut scorer = FilteredScorer::new(raw_scorer.as_ref(), Some(&fake_filter_context));
+            let mut scorer = vector_holder.get_scorer(query);
 
             let mut points_to_score = (0..1500)
                 .map(|_| rng.random_range(0..NUM_VECTORS) as u32)
                 .collect_vec();
-            scorer.score_points(&mut points_to_score, 1000);
+            scorer.score_points(&mut points_to_score, 1000).black_box();
         })
     });
 
@@ -94,15 +88,15 @@ fn scoring_vectors(c: &mut Criterion) {
 
     group.bench_function("score-point", |b| {
         b.iter(|| {
-            let fake_filter_context = FakeFilterContext {};
             let query = random_vector(&mut rng, DIM);
-            let raw_scorer = vector_holder.get_raw_scorer(query).unwrap();
-            let mut scorer = FilteredScorer::new(raw_scorer.as_ref(), Some(&fake_filter_context));
+            let mut scorer = vector_holder.get_scorer(query);
 
             let mut points_to_score = (0..points_per_cycle)
                 .map(|_| rng.random_range(0..num_vectors) as u32)
                 .collect_vec();
-            scorer.score_points(&mut points_to_score, points_per_cycle);
+            scorer
+                .score_points(&mut points_to_score, points_per_cycle)
+                .black_box();
         })
     });
 
@@ -111,15 +105,15 @@ fn scoring_vectors(c: &mut Criterion) {
 
     group.bench_function("score-point-10x", |b| {
         b.iter(|| {
-            let fake_filter_context = FakeFilterContext {};
             let query = random_vector(&mut rng, DIM);
-            let raw_scorer = vector_holder.get_raw_scorer(query).unwrap();
-            let mut scorer = FilteredScorer::new(raw_scorer.as_ref(), Some(&fake_filter_context));
+            let mut scorer = vector_holder.get_scorer(query);
 
             let mut points_to_score = (0..points_per_cycle)
                 .map(|_| rng.random_range(0..num_vectors) as u32)
                 .collect_vec();
-            scorer.score_points(&mut points_to_score, points_per_cycle);
+            scorer
+                .score_points(&mut points_to_score, points_per_cycle)
+                .black_box();
         })
     });
 
@@ -128,15 +122,15 @@ fn scoring_vectors(c: &mut Criterion) {
 
     group.bench_function("score-point-50x", |b| {
         b.iter(|| {
-            let fake_filter_context = FakeFilterContext {};
             let query = random_vector(&mut rng, DIM);
-            let raw_scorer = vector_holder.get_raw_scorer(query).unwrap();
-            let mut scorer = FilteredScorer::new(raw_scorer.as_ref(), Some(&fake_filter_context));
+            let mut scorer = vector_holder.get_scorer(query);
 
             let mut points_to_score = (0..points_per_cycle)
                 .map(|_| rng.random_range(0..num_vectors) as u32)
                 .collect_vec();
-            scorer.score_points(&mut points_to_score, points_per_cycle);
+            scorer
+                .score_points(&mut points_to_score, points_per_cycle)
+                .black_box();
         })
     });
 }