--- a/lib/segment/benches/hnsw_build_graph.rs
+++ b/lib/segment/benches/hnsw_build_graph.rs
@@ -5,9 +5,8 @@ use common::types::PointOffsetType;
 use criterion::{Criterion, criterion_group, criterion_main};
 use rand::SeedableRng;
 use rand::rngs::StdRng;
-use segment::fixtures::index_fixtures::{FakeFilterContext, TestRawScorerProducer};
+use segment::fixtures::index_fixtures::TestRawScorerProducer;
 use segment::index::hnsw_index::graph_layers_builder::GraphLayersBuilder;
-use segment::index::hnsw_index::point_scorer::FilteredScorer;
 use segment::spaces::simple::CosineMetric;
 use segment::vector_storage::chunked_vector_storage::VectorOffsetType;
 
@@ -27,11 +26,9 @@ fn hnsw_benchmark(c: &mut Criterion) {
             let mut rng = rand::rng();
             let mut graph_layers_builder =
                 GraphLayersBuilder::new(NUM_VECTORS, M, M * 2, EF_CONSTRUCT, 10, USE_HEURISTIC);
-            let fake_filter_context = FakeFilterContext {};
             for idx in 0..(NUM_VECTORS as PointOffsetType) {
                 let added_vector = vector_holder.vectors.get(idx as VectorOffsetType).to_vec();
-                let raw_scorer = vector_holder.get_raw_scorer(added_vector).unwrap();
-                let scorer = FilteredScorer::new(raw_scorer.as_ref(), Some(&fake_filter_context));
+                let scorer = vector_holder.get_scorer(added_vector);
                 let level = graph_layers_builder.get_random_layer(&mut rng);
                 graph_layers_builder.set_levels(idx, level);
                 graph_layers_builder.link_new_point(idx, scorer);