@generall 2025-03-25
After review in-call:

- Try to merge Filterer and RawScorer in the following way:

FIlteredScorer uses RawScorer to combine into a single call
```
            filterer.filter_scores(&mut scores, limit);
            scorer.score_points(&mut scores);
```

If RawScorer is needed without filters, it can be extracted (used from) FilteredScorer


- `is_stopped` flag should be moved FilteredScorer
- Move `peek_top_iter` and `peek_top_all` into FilteredScorer


---

For a separate PR:

- Try to use `query_scorer.score_stored_batch` in `score_points` of RawScorerImpl